requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
gunicorn==21.2.0
boto3==1.28.85
python-dotenv==1.0.0
marshmallow==3.20.1
pytest==7.4.2
pytest-flask==1.2.0
psutil==5.9.6
